Fixes for string

+++ string.h
@@ -1,7 +1,7 @@
 #ifndef __STRING_H__
 #define __STRING_H__
 
-typedef unsigned long size_t;
+typedef int size_t;
 
 #include "kmem.h"
